index/pyproject.toml
[tool.poetry]
name = "tai"
version = "0.1.0"
description = ""
authors = ["William Goodall <wgoodall01@gmail.com>"]
readme = "README.md"
packages = [{include = "tai_index"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
ipykernel = "^6.25.2"
llama-index = "^0.8.45.post1"
pypdf = "^3.16.4"
docx2txt = "^0.8"
torch = "^2.1.0"
transformers = "^4.34.0"
python-pptx = "^0.6.22"
jupyter = "^1.0.0"
ipywidgets = "^8.1.1"
tqdm = "^4.66.1"
pillow = ">=3.3.2,<=9.5.0"
python-dotenv = "^1.0.0"
pymongo = "^4.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
